Move all EnableHA functions out of addmachine.go
Instead of clumping them all together in state_test.go and
addmachine.go, pull them out into their own file enableha.go and
enableha_test.go. Should be more understandable when they are grouped
together like that.
